Move search into bottom-tab stacks, add unmountOnBlur
