+++ packages/mobile/src/components/app-navigator/BaseStackNavigator.tsx
@@ -6,6 +6,8 @@ import { MessageType } from 'audius-client/src/services/native-mobile-interface/
 
 import { useDispatchWeb } from 'app/hooks/useDispatchWeb'
 import { ProfileScreen } from 'app/screens/profile-screen'
+import { SearchResultsScreen } from 'app/screens/search-results-screen'
+import { SearchScreen } from 'app/screens/search-screen'
 import { TrackScreen } from 'app/screens/track-screen'
 import {
   FavoritedScreen,
@@ -15,6 +17,12 @@ import {
 } from 'app/screens/user-list-screen'
 
 import { TopBar } from './TopBar'
+
+const forFade = ({ current }) => ({
+  cardStyle: {
+    opacity: current.progress
+  }
+})
 
 type BaseNavigatorProps = {
   baseScreen: (
@@ -58,6 +66,14 @@ export const BaseStackNavigator = ({
       {baseScreen(Stack)}
       <Stack.Screen name='track' component={TrackScreen} />
       <Stack.Screen name='profile' component={ProfileScreen} />
+      <Stack.Group>
+        <Stack.Screen
+          name='Search'
+          component={SearchScreen}
+          options={{ cardStyleInterpolator: forFade }}
+        />
+        <Stack.Screen name='SearchResults' component={SearchResultsScreen} />
+      </Stack.Group>
       <Stack.Group>
         <Stack.Screen name='FollowersScreen' component={FollowersScreen} />
         <Stack.Screen name='FollowingScreen' component={FollowingScreen} />
